refactor(api): consolidate HTTP response types into shared/; move controllers out of Kreditozrouti/
- Delete api CoursesResponse, StudyPlansResponse, StudyPlanCoursesResponse types — shared/http/responses.ts is now the single source of truth for response shapes
- Update CoursesController, StudyPlansController, StudyPlanCoursesController to use plain Response (no generic) — removes duplicate type annotation that mirrored the shared DTO
- Update CourseCacheService to import CoursesFilter from @shared/http/courses and CoursesResponseDTO['facets'] from @shared/http/responses (was importing from controller re-export and local type)
- Update StudyPlanCacheService same way for StudyPlansFilter + StudyPlansResponseDTO['facets']
- Remove stale "must stay in sync" comment from shared/http/responses.ts — no longer needed since there are no parallel api types to sync with
- Rename Controllers/Kreditozrouti/ → Controllers/Courses/ + Controllers/StudyPlans/ for cleaner domain organisation
- Move api/src/utils/ical.ts → api/src/Services/ICalService.ts to match service naming convention
- Delete CoursesRequest, StudyPlansRequest, StudyPlanCoursesRequest type aliases (zero-value wrappers around shared filter types)
