--- api/src/Controllers/Kreditozrouti/types/CoursesRequest.ts
@@ -1,7 +1,0 @@
-import { CoursesFilter } from '@api/Controllers/Kreditozrouti/CoursesController'
-
-/**
- * Request type for the courses endpoint.
- */
-type CoursesRequest = CoursesFilter
-export default CoursesRequest
