--- api/src/Controllers/Kreditozrouti/types/StudyPlanCoursesRequest.ts
@@ -1,7 +1,0 @@
-import { StudyPlanCoursesFilter } from '@api/Controllers/Kreditozrouti/StudyPlanCoursesController'
-
-/**
- * Request type for the study plan courses endpoint.
- */
-type StudyPlanCoursesRequest = StudyPlanCoursesFilter
-export default StudyPlanCoursesRequest
