--- api/src/Controllers/Kreditozrouti/types/StudyPlansRequest.ts
@@ -1,7 +1,0 @@
-import { StudyPlansFilter } from '@api/Controllers/Kreditozrouti/StudyPlansController'
-
-/**
- * Request type for the study plans endpoint.
- */
-type StudyPlansRequest = StudyPlansFilter
-export default StudyPlansRequest
