+++ api/src/Services/StudyPlan/StudyPlanFacetService.ts
@@ -1,6 +1,6 @@
 import type { FacetItem } from '@shared/http/facets'
 import { mysql } from '@api/clients'
-import { StudyPlansFilter } from '@api/Controllers/Kreditozrouti/StudyPlansController'
+import { StudyPlansFilter } from '@api/Controllers/StudyPlans/StudyPlansController'
 import { ExcludeMethods, StudyPlan, StudyPlanTable } from '@api/Database/types'
 import { StudyPlanCacheService } from './StudyPlanCacheService'
 import { StudyPlanFilterBuilder } from './StudyPlanFilterBuilder'
